Restructure project into package layout
