+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+## [0.1.0] - 2024-05-01
+### Added
+- Initial project restructuring introducing the `neva` package with dedicated modules for agents, environments, schedulers, tools, memory, and utilities.
