+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,9 @@
+# Code of Conduct
+
+We are committed to fostering a welcoming and inclusive community. All participants are expected to:
+
+- Be respectful and considerate in all communications.
+- Provide constructive feedback and accept it graciously.
+- Uphold professional conduct and refrain from harassment or discrimination of any kind.
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by contacting the project maintainers at `maintainers@example.com`.
