+++ neva/__init__.py
@@ -1,0 +1,12 @@
+"""Neva core package providing agents, environments, and schedulers."""
+
+from . import agents, environments, memory, schedulers, tools, utils
+
+__all__ = [
+    "agents",
+    "environments",
+    "memory",
+    "schedulers",
+    "tools",
+    "utils",
+]
