+++ neva/tools/__init__.py
@@ -1,0 +1,215 @@
+"""Tool implementations used by Neva agents."""
+
+from __future__ import annotations
+
+import ast
+import operator
+from typing import Any, Callable, Dict, Optional, Protocol, cast
+
+import logging
+
+from neva.agents.base import Tool
+from neva.utils.exceptions import MissingDependencyError, ToolExecutionError
+
+try:  # pragma: no cover - exercised when wikipedia is available.
+    import wikipedia  # type: ignore
+except Exception:  # pragma: no cover - fallback for optional dependency.
+    wikipedia = None  # type: ignore
+
+
+logger = logging.getLogger(__name__)
+
+
+def _missing_dependency_message(package: str, *, fallback: str) -> str:
+    return (
+        f"The optional dependency `{package}` is unavailable. {fallback} "
+        f"Install it with `pip install {package}` or provide a custom factory "
+        "when constructing the tool to continue running experiments without "
+        "network access."
+    )
+
+
+_ALLOWED_OPERATORS: Dict[type, Callable[[float, float], float]] = {
+    ast.Add: operator.add,
+    ast.Sub: operator.sub,
+    ast.Mult: operator.mul,
+    ast.Div: operator.truediv,
+    ast.Pow: operator.pow,
+    ast.Mod: operator.mod,
+}
+
+
+class TranslatorBackend(Protocol):
+    """Protocol describing the translator interface used by :class:`TranslatorTool`."""
+
+    def translate(self, text: str, dest: str) -> Any:
+        ...
+
+
+class MathTool(Tool):
+    """Safely evaluate basic mathematical expressions."""
+
+    def __init__(self) -> None:
+        super().__init__(
+            "calculator",
+            "performs arithmetic expressions",
+            capabilities=["math", "calculation"],
+        )
+
+    def _eval_node(self, node: ast.AST) -> float:
+        if isinstance(node, ast.Num):  # pragma: no cover - python<3.8 fallback
+            return float(node.n)  # type: ignore[attr-defined]
+        if isinstance(node, ast.Constant) and isinstance(node.value, (int, float)):
+            return float(node.value)
+        if isinstance(node, ast.BinOp):
+            op_type = type(node.op)
+            if op_type not in _ALLOWED_OPERATORS:
+                raise ToolExecutionError(
+                    f"Unsupported operator: {op_type.__name__}"
+                )
+            left = self._eval_node(node.left)
+            right = self._eval_node(node.right)
+            return _ALLOWED_OPERATORS[op_type](left, right)
+        if isinstance(node, ast.UnaryOp) and isinstance(node.op, (ast.UAdd, ast.USub)):
+            value = self._eval_node(node.operand)
+            return value if isinstance(node.op, ast.UAdd) else -value
+        raise ToolExecutionError("Unsupported expression")
+
+    def use(self, task: str) -> str:
+        try:
+            expression = ast.parse(task, mode="eval")
+            result = self._eval_node(expression.body)
+            return str(result)
+        except ToolExecutionError:
+            raise
+        except Exception as exc:
+            raise ToolExecutionError(f"Failed to use MathTool: {exc}") from exc
+
+
+class WikipediaTool(Tool):
+    """Retrieve short summaries from Wikipedia."""
+
+    def __init__(self, *, summary_sentences: int = 2) -> None:
+        super().__init__(
+            "wikipedia",
+            "looks up concise encyclopedia summaries",
+            capabilities=["search", "encyclopedia"],
+        )
+        self.summary_sentences = summary_sentences
+
+    def use(self, task: str) -> str:
+        if wikipedia is None:
+            raise MissingDependencyError(
+                _missing_dependency_message(
+                    "wikipedia",
+                    fallback="Wikipedia lookups require the `wikipedia` package.",
+                )
+            )
+
+        try:
+            return wikipedia.summary(task, sentences=self.summary_sentences)
+        except Exception as exc:
+            logger.warning("Wikipedia lookup failed: %s", exc)
+            raise ToolExecutionError(
+                "Failed to use WikipediaTool because the external service "
+                f"returned an error: {exc}. Provide a stub via"
+                " `summary_sentences` or monkeypatch the backend during tests."
+            ) from exc
+
+
+class TranslatorTool(Tool):
+    """Translate text using a configurable backend."""
+
+    def __init__(
+        self,
+        *,
+        translator_factory: Optional[Callable[[], object]] = None,
+        target_language: str = "en",
+    ) -> None:
+        super().__init__(
+            "translator",
+            "translates text to different languages",
+            capabilities=["translation"],
+        )
+        self._translator_factory = translator_factory
+        self.target_language = target_language
+
+    def _get_translator(self) -> TranslatorBackend:
+        if self._translator_factory is not None:
+            translator = self._translator_factory()
+            return cast(TranslatorBackend, translator)
+
+        try:  # pragma: no cover - depends on optional dependency.
+            from googletrans import Translator  # type: ignore
+        except Exception as exc:  # pragma: no cover - optional dependency missing.
+            raise MissingDependencyError(
+                _missing_dependency_message(
+                    "googletrans",
+                    fallback="Translation requires access to the Google Translate API.",
+                )
+            ) from exc
+
+        return cast(TranslatorBackend, Translator())
+
+    def use(self, task: str) -> str:
+        try:
+            translator: TranslatorBackend = self._get_translator()
+            result: Any = translator.translate(task, dest=self.target_language)
+            return getattr(result, "text", str(result))
+        except Exception as exc:
+            logger.warning("Translator backend failed: %s", exc)
+            raise ToolExecutionError(
+                "Failed to use TranslatorTool because the translation backend "
+                f"raised an error: {exc}. Provide a stub translator factory for"
+                " deterministic tests."
+            ) from exc
+
+
+class SummarizerTool(Tool):
+    """Summarise text using a pluggable backend."""
+
+    def __init__(
+        self,
+        *,
+        summarizer_factory: Optional[Callable[[], Callable[[str], str]]] = None,
+    ) -> None:
+        super().__init__(
+            "summarizer",
+            "creates short summaries of longer documents",
+            capabilities=["summarisation"],
+        )
+        self._summarizer_factory = summarizer_factory
+
+    def _get_summarizer(self) -> Callable[[str], str]:
+        if self._summarizer_factory is not None:
+            summarizer = self._summarizer_factory()
+            return cast(Callable[[str], str], summarizer)
+
+        try:  # pragma: no cover - optional dependency.
+            from summarizer import Summarizer  # type: ignore
+        except Exception as exc:  # pragma: no cover
+            raise MissingDependencyError(
+                _missing_dependency_message(
+                    "summarizer",
+                    fallback="Summarisation requires the `bert-extractive-summarizer` package.",
+                )
+            ) from exc
+
+        model = Summarizer()
+
+        def _summarise(text: str, *, _model=model) -> str:
+            return _model(text, min_length=60, max_length=100)
+
+        return _summarise
+
+    def use(self, task: str) -> str:
+        try:
+            summariser = self._get_summarizer()
+            return summariser(task)
+        except Exception as exc:
+            logger.warning("Summariser backend failed: %s", exc)
+            raise ToolExecutionError(
+                "Failed to use SummarizerTool because the summarisation backend "
+                f"raised an error: {exc}. Pass a lightweight factory during"
+                " offline experimentation."
+            ) from exc
