+++ pyproject.toml
@@ -1,0 +1,11 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "neva"
+version = "0.1.0"
+description = "Neva provides agent, environment, and scheduler abstractions for LLM-based simulations."
+readme = "README.md"
+requires-python = ">=3.8"
+authors = [{name = "Neva Contributors"}]
